# By Joey Kleingers (4) and Michael Groeber (1)
# Via Michael Groeber
* 'develop' of ssh://scm.bluequartz.net/DREAM3D:
  updating gbcd to use doubles instead of float
  Directories and files are now being added to the prebuilts list using the new recursive algorithm.
  The directories are now created recursively, but the pipeline items are not.
  Continued work on recursive algorithm
  In the process of writing a recursive function to display any structure under the Prebuilt section

+++ Source/DREAM3DLib/StatisticsFilters/FindGBCD.h
@@ -141,7 +141,7 @@ class DREAM3DLib_EXPORT FindGBCD : public SurfaceMeshFilter
     float* m_FieldEulerAngles;
     int32_t* m_FieldPhases;
     unsigned int* m_CrystalStructures;
-    float* m_GBCD;
+    double* m_GBCD;
 
     FindGBCD(const FindGBCD&); // Copy Constructor Not Implemented
     void operator=(const FindGBCD&); // Operator '=' Not Implemented
